Add Windows compatibility
Unfortunately the embedded browser did not work
with elevated mode, so it was required to keep the
"sudo" part. In Windows there is no easy way to do
it, but there are many "sudo" implementations, so
this way the source code can be unchanged and
the user can select for themselves any sudo
implementation.